+++ CodeRed/Interface/GpuPipelineState/GpuRasterizationState.hpp
@@ -13,7 +13,7 @@ namespace CodeRed {
 	protected:
 		explicit GpuRasterizationState(
 			const std::shared_ptr<GpuLogicalDevice>& device,
-			const FrontFace front_face = FrontFace::Clockwise,
+			const FrontFace front_face = FrontFace::CounterClockwise,
 			const CullMode cull_mode = CullMode::Back,
 			const FillMode fill_mode = FillMode::Solid,
 			const bool depth_clamp = true) :