+++ CodeRed/Shared/Information/TextureRefInfo.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "../Enum/TextureRefUsage.hpp"
+#include "../Enum/PixelFormat.hpp"
 
 #include "../ValueRange.hpp"
 
@@ -11,17 +12,19 @@ namespace CodeRed {
 		ValueRange<size_t> Array = ValueRange<size_t>(0, 0);
 
 		TextureRefUsage Usage = TextureRefUsage::Common;
+		PixelFormat Format = PixelFormat::Unknown;
 
 		TextureRefInfo() = default;
 
-		TextureRefInfo(const TextureRefUsage& usage) :
-			Usage(usage) {}
+		TextureRefInfo(const TextureRefUsage& usage, const PixelFormat& format = PixelFormat::Unknown) :
+			Usage(usage), Format(format) {}
 		
 		TextureRefInfo(
 			const ValueRange<size_t>& mipLevel,
 			const ValueRange<size_t>& array,
-			const TextureRefUsage& usage = TextureRefUsage::Common) :
-			MipLevel(mipLevel), Array(array), Usage(usage) {}
+			const TextureRefUsage& usage = TextureRefUsage::Common,
+			const PixelFormat& format = PixelFormat::Unknown) :
+			MipLevel(mipLevel), Array(array), Usage(usage), Format(format) {}
 	};
 	
 }